+++ example/lib/main.dart
@@ -4,16 +4,20 @@ import 'package:flutter/material.dart';
 void main() => runApp(~MyApp());
 
 class MyApp extends Application {
-  final counterState = 0.state;
+  final _counter$ = 0.$;
+
+  int get counter => _counter$.value;
+
+  set counter(int v) => _counter$.value = v;
 
   @override
   Widget render(BuildContext context) => Scaffold(
         appBar: AppBar(
           title: const Text('Flutter Demo Home Page'),
         ),
-        body: Counter(counterState),
+        body: Counter(_counter$),
         floatingActionButton: FloatingActionButton(
-          onPressed: () => counterState.value++,
+          onPressed: () => counter++,
           tooltip: 'Increment',
           child: const Icon(Icons.add),
         ),
@@ -22,6 +26,10 @@ class MyApp extends Application {
 
 class Counter extends Component<int> {
   const Counter(super.state, {super.key});
+
+  int get _value => state.value;
+
+  set _value(int v) => state.value = v;
 
   @override
   Widget render(BuildContext context) {
@@ -33,11 +41,11 @@ class Counter extends Component<int> {
             'You have pushed the button this many times:',
           ),
           Text(
-            '${state.value}',
+            '$_value',
             style: Theme.of(context).textTheme.headlineMedium,
           ),
           ElevatedButton(
-            onPressed: () => state.value = 0,
+            onPressed: () => _value = 0,
             child: const Text('Reset'),
           ),
         ],