+++ other_examples/shopping_cart/lib/components/fab.dart
@@ -5,9 +5,11 @@ import 'package:shopping_cart/models/item.dart';
 class Fab extends Component<bool> {
   const Fab(super.state, {super.key});
 
+  bool get _visible => state.value;
+
   @override
   Widget render(BuildContext context) => Visibility(
-      visible: state.value,
+      visible: _visible,
       child: FloatingActionButton.extended(
         onPressed: () => context.url = '/cart',
         label: Text('${Item.cart.length} cart'),