+++ other_examples/todo_list/lib/components/header_text.dart
@@ -1,0 +1,9 @@
+import 'package:esm/esm.dart';
+import 'package:flutter/material.dart';
+
+class HeaderText extends Component<String> {
+  const HeaderText(super.state, {super.key});
+
+  @override
+  Widget render(BuildContext context) => Text(state.value);
+}