+++ other_examples/todo_list/lib/main.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+import 'package:todo_list/todo_app.dart';
+
+void main() => runApp(~TodoApp());