+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
-* First version.
+- First version.
+
+## 1.1.0
+
+- Change state builder extension name function to $.
+- Add optional VoidCallback parameter to state update function.
+- Update example projects.